We now use pngs instead of tgas in order to:    A) Utilize an existing and most likely more efficient library    B) Have much smaller file sizes
Textures are still a mess, though. There is an additional "reticule" that looks like garbage that is strangely hidden behind blackness. Not sure what that is.

The good news is that the Arwing now controls MUCH more smoothly. That's something, right?

Anyone who wishes to test this out will need to download the new data folder from the downloads section.