Some major changed behind the scenes. Still having trouble with the asteroids, so they have been temporarily replaced with blue boxes (that flicker, sadly). You can fire a laser with the "1" button, but it will not follow the path that you fire it in.
TODO: The asteroids, when calling the display list for that model, will give a DSI error (which is why they are replaced with blue boxes). The trouble does indeed come from the display list, for if the blue-box code is placed within the display list generation for the models, it gives a DSI error.

What is puzzling, however, is the fact that the ARwing and the laser are both generating display lists just fine. The ARwing still has trippy textures.